complete problem0

+++ pset0/bc61.cc
@@ -1,11 +1,10 @@
-#include <cstdlib>
 #include <cstdio>
 
 int main() {
-    int cnt = 0;
+    unsigned long cnt = 0;
     while (fgetc(stdin) != EOF) {
         cnt++;
     }
-    fprintf(stdout, "%d\n", cnt);
+    fprintf(stdout, "%lu\n", cnt);
     exit(0);
 }
